Conversion

Description

This module is responsible for putting everthing together: It implements algorithms for transforming a Solid into it's pieces and pieces into OpenSCAD source.

Documentation

stlToOpenSCAD :: String -> String

This function takes as input a String of stl source and returns OpenSCAD source for the pieces relative to that stl file – in the form of a String