Geometry

Description

Here are implemented types relative to geometric notions: Points, Faces and Solids, besides functions involving Points (which are really vectors).

Types

data Point

A point in 3D Euclidian Space

Constructors

Point Float Float Float

data Face

A face is defined by a normal and three vertices

Constructors

Face

Fields

normal, p, q, r :: Point

data Solid

A Solid is a list of Faces

Constructors

Solid [Face]

Point Functions

angle :: Point -> Point -> Float

Returns angle between two Points

dot :: Point -> Point -> Float

Dot product of two Points

norm :: Point -> Float

Norm of a Point

distance :: Point -> Point -> Float

Euclidian distance between two Points